@@ -7,19 +7,36 @@ publish_to: 'none'
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.17.0 <3.0.0"
+  sdk: '>=2.18.0-165.1.beta <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
   flutter_localizations:
     sdk: flutter
+  google_fonts: ^3.0.1
+  flutter_bloc: ^8.0.1
+  bloc:
+  fluentui_system_icons: ^1.1.176
+  shared_preferences: ^2.0.15
+  animations: ^2.0.3
+  dio: ^4.0.6
+  equatable:
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
 
   flutter_lints: ^2.0.0
+  # package for setting up the app icon
+  flutter_launcher_icons: "^0.9.2"
+
+# flutter pub get
+# flutter pub run flutter_launcher_icons:main 
+flutter_icons:
+  android: "launcher_icon"
+  ios: true
+  image_path: "assets/icon/icon.png"
 
 flutter:
   uses-material-design: true
@@ -30,3 +47,6 @@ flutter:
   assets:
     # Add assets from the images directory to the application.
     - assets/images/
+    - assets/images/SVG/
+    - assets/images/1000w/
+    - assets/images/200w/
