@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 
 /// Displays detailed information about a SampleItem.
 class SampleItemDetailsView extends StatelessWidget {
-  const SampleItemDetailsView({Key? key}) : super(key: key);
+  const SampleItemDetailsView({super.key});
 
   static const routeName = '/sample_item';
 
