@@ -1,6 +1,6 @@
 import type { HttpContextContract } from '@ioc:Adonis/Core/HttpContext'
-import Account, { AccountType } from 'App/Models/accounts/Account'
-import CreateAccountValidator, { DestroyAccountValidator, ListAccountsValidator, ShowAccountValidator, UpdateAccountValidator } from 'App/Validators/AccountValidator'
+import Account, { AccountType } from 'App/Models/Account'
+import  { CreateAccountValidator, DestroyAccountValidator, ListAccountsValidator, ShowAccountValidator, UpdateAccountValidator } from 'App/Validators/AccountValidator'
 import { Image } from 'App/Models/File'
 import { ModelObject } from '@ioc:Adonis/Lucid/Orm';
 
@@ -13,18 +13,26 @@ export default class AccountsController {
     * @example
     * curl -X GET -H "Content-Type: application/json" -d '{"page": 1}' http://localhost:3333/api/v1/accounts
     */
-  public async index({ auth, request }: HttpContextContract): Promise<{
+  public async index({ auth, request, bouncer }: HttpContextContract): Promise<{
     meta: any;
     data: ModelObject[];
   }> {
     const payload = await request.validate(ListAccountsValidator)
-    var accountsQuery = auth.user!.related("accounts").query()
+    await bouncer.with('AccountPolicy').authorize('viewList', payload)
+    var accountsQuery = Account.query()
     var page = 1
     var limit = 24
 
-    // if (payload.query) {
+
     if (payload.search) {
-      accountsQuery = accountsQuery.where('name', 'like', `%${payload.search}%`)
+      for (let i = 0; i < payload.search_by!.length; i++) {
+        const element = payload.search_by![i];
+        if (i == 0) {
+          accountsQuery = accountsQuery.where(element, 'like', `%${payload.search}%`)
+        } else {
+          accountsQuery = accountsQuery.orWhere(element, 'like', `%${payload.search}%`)
+        }
+      }
     }
     if (payload.sort) {
       accountsQuery = accountsQuery.orderBy(payload.sort, payload.order)
@@ -34,35 +42,41 @@ export default class AccountsController {
         accountsQuery = accountsQuery.preload(load)
       }
     }
+    if (payload.where) {
+      for (const key in payload.where) {
+        accountsQuery = accountsQuery.where(key, payload.where[key])
+      }
+    }
     page = payload.page || page
     limit = payload.limit || limit
 
-    // }
     return (await accountsQuery.paginate(page, Math.min(limit, 24))).toJSON()
   }
 
   /**
    * creates a new account for the user
    * @param {HttpContextContract}
-   * @returns {Promise<{ account: ModelObject; avatar: Image | null; }>}
+   * @returns {Promise<{ account: ModelObject; photo: Image | null; }>}
    * @memberof AccountsController
    * @example
    * curl -X PUT -H "Content-Type: application/json" -d '{"name": "My Account", "type": "Bank", "number": "123456789"}' http://localhost:3333/api/v1/accounts
    */
-  public async store({ request, auth }: HttpContextContract): Promise<{ account: ModelObject; avatar: Image | null; }> {
+  public async store({ request, auth, bouncer }: HttpContextContract): Promise<{ account: ModelObject; photo: Image | null; }> {
+
+    await bouncer.with('AccountPolicy').authorize('create', null)
     const payload = await request.validate(CreateAccountValidator)
     const account = await auth.user!.related("accounts").create({
       name: payload.name,
       description: payload.description,
       type: payload.type as AccountType,
     })
-    var avatar: Image | null = null;
-    if (payload.avatar) {
-      avatar = await account.setAvatar(payload.avatar)
+    var photo: Image | null = null;
+    if (payload.photo) {
+      photo = await account.setPhoto(payload.photo)
     }
     return {
       account: account.toJSON(),
-      avatar: avatar,
+      photo: photo,
     }
   }
 
@@ -74,10 +88,18 @@ export default class AccountsController {
    * @example
    * curl -X GET -H "Content-Type: application/json" http://localhost:3333/api/v1/accounts/1
    */
-  public async show({ auth, request }: HttpContextContract) {
+  public async show({ auth, request, bouncer }: HttpContextContract): Promise<any> {
     const payload = await request.validate(ShowAccountValidator)
-    var account = await Account.query().where('id', payload.params.id).where('user_id', auth.user!.id).first()
-    await account!.load('avatar')
+    var account = await Account.query().where('id', payload.params.id).first()
+    await bouncer.with('AccountPolicy').authorize('view', account)
+    if (payload.load) {
+      for (const load of payload.load) {
+        await account!.load(load)
+      }
+    }
+    if (!payload.load?.includes('photos')) {
+      await account!.load('photos')
+    }
     return account!.toJSON()
   }
 
@@ -89,20 +111,21 @@ export default class AccountsController {
    * @example
    * curl -X PUT -H "Content-Type: application/json" -d '{"name": "My Account", "type": "Bank", "number": "123456789"}' http://localhost:3333/api/v1/accounts/1
    */
-  public async update({ request }: HttpContextContract): Promise<any> {
+  public async update({ request, bouncer }: HttpContextContract): Promise<any> {
     // validate also params.id
     const payload = await request.validate(UpdateAccountValidator)
     const account = (await Account.find(payload.params.id))!
+    await bouncer.with('AccountPolicy').authorize('update', account)
     account.name = payload.name ?? account.name
     account.description = payload.description ?? account.description
     await account.save()
-    var avatar: Image | null = null;
-    if (payload.avatar) {
-      await account.setAvatar(payload.avatar)
+    var photo: Image | null = null;
+    if (payload.photo) {
+      await account.setPhoto(payload.photo)
     }
     return {
       account: account.toJSON(),
-      avatar: avatar,
+      photo: photo,
     }
   }
 
@@ -114,10 +137,11 @@ export default class AccountsController {
    * @example
    * curl -X DELETE -H "Content-Type: application/json" http://localhost:3333/api/v1/accounts/{accountId}
    */
-  public async destroy({ request }: HttpContextContract): Promise<void> {
+  public async destroy({ request, bouncer }: HttpContextContract): Promise<void> {
     const payload = await request.validate(DestroyAccountValidator)
-    const account = await Account.find(payload.params.id)
-    return await account!.delete()
+    const account = (await Account.find(payload.params.id))!
+    await bouncer.with('AccountPolicy').authorize('delete', account)
+    return await account.delete()
   }
 
 
