@@ -1,40 +1,52 @@
-import {  belongsTo, BelongsTo, column } from '@ioc:Adonis/Lucid/Orm'
+import {  belongsTo, BelongsTo, column, manyToMany, ManyToMany } from '@ioc:Adonis/Lucid/Orm'
 import { DateTime } from 'luxon'
+import DotBaseModel from 'Dot/models/DotBaseModel'
+import { usedPivot } from 'Dot/hooks/orm'
 import User from '../User'
-import DotBaseModel from 'Dot/models/DorBaseModel'
+import { string } from '@ioc:Adonis/Core/Helpers'
 
 export default class Phone extends DotBaseModel {
 
-  @column({ isPrimary: true })
-  public type: string
+  // @column()
+  // public code: number
 
-  @column({ isPrimary: true })
+  @column()
   public value: string
 
   @column.dateTime()
   public validatedAt: DateTime
 
-  // @column.dateTime({ autoCreate: true })
-  // public createdAt: DateTime
+  @column.dateTime({ autoCreate: true })
+  public createdAt: DateTime
 
-  // @column.dateTime({ autoCreate: true, autoUpdate: true })
-  // public updatedAt: DateTime
-  //
+  @column.dateTime({ autoCreate: true, autoUpdate: true })
+  public updatedAt: DateTime
+
   @column({ serializeAs: null })
-  public relatedTo: string
+  public relatedId: string
 
   @column({ serializeAs: null })
   public relatedType: string
 
-  @belongsTo(() => User, { foreignKey: 'relatedTo' })
-  public user: BelongsTo<typeof User>
+  @manyToMany(()=>User, {
+    pivotForeignKey: 'phone_id',
+    pivotRelatedForeignKey: 'related_id',
+    pivotTable: "phones_pivot",
+    pivotColumns: ['tag'],
+    onQuery: (builder) => {
+      // if (tag) {
+      //   builder.wherePivot('tag', tag);
+      // }
+    }
+  })
+  public users: ManyToMany<typeof User>
 
   //
   // @column()
   // public userId: string
 
   // belongsTo returns the [User] that owns the [ContactOption]
-  // @belongsTo(() => DotBaseModel,{foreignKey:"id", localKey:"relatedTo"})
+  // @belongsTo(() => DotBaseModel,{foreignKey:"id", localKey:"relatedId"})
   // public owner: BelongsTo<typeof DotBaseModel>
 
   // static method to test if the given phone is valid
