@@ -1,31 +1,18 @@
-import { DateTime } from 'luxon'
-import { BaseModel, belongsTo, BelongsTo,  column,  hasMany,  HasMany,  HasOne, hasOne, LucidModel } from '@ioc:Adonis/Lucid/Orm'
 import Email from '../../ContactOptions/Email'
 import Phone from '../../ContactOptions/Phone'
-import Address from '../../Address'
-import DotBaseModel from 'Dot/models/DorBaseModel'
+import DotBaseModel from 'Dot/models/DotBaseModel'
+import { column, hasMany, HasMany } from '@ioc:Adonis/Lucid/Orm'
 
 export enum HumanGender {
     male = 0,
     female = 1,
 }
 export class ProfileModel extends DotBaseModel  {
-    // @column.dateTime({ autoCreate: true })
-    // public createdAt: DateTime
-    // @column.dateTime({ autoCreate: true, autoUpdate: true })
-    // public updatedAt: DateTime
-    @column()
-    public relatedTo: string
-    // has address
-    @hasMany(() => Email, { foreignKey: 'relatedTo' })
-    public emails: HasMany<typeof Email>
-    @hasMany(() => Phone, { foreignKey: 'relatedTo' })
-    public phones: HasMany<typeof Phone>
 }
 
 
 // export default class Project extends compose(DotBaseModel, ProfileMixin) {
-    
+
 // }
 
 
@@ -34,7 +21,7 @@ export class ProfileModel extends DotBaseModel  {
 //     public gender: HumanGender
 //     public createdAt: DateTime
 //     public updatedAt: DateTime
-//     public relatedTo: string
+//     public relatedId: string
 //     public address: HasOne<typeof Address>
 //     public email: HasOne<typeof Email>
 //     public phone: HasOne<typeof Phone>
