@@ -2,13 +2,13 @@ import { schema, rules } from '@ioc:Adonis/Core/Validator'
 import type { HttpContextContract } from '@ioc:Adonis/Core/HttpContext'
 import DotValidator from './DotValidator'
 
-export default class CreateMerchantProfileValidator implements DotValidator {
+export class CreateMerchantProfileValidator implements DotValidator {
   constructor(protected ctx: HttpContextContract) {
   }
   public schema = schema.create({
     account_id: schema.string({}, [
       rules.required(),
-      rules.exists({ table: 'accounts', column: 'id'}),
+      rules.exists({ table: 'accounts', column: 'id' }),
     ]),
 
   })
@@ -16,8 +16,8 @@ export default class CreateMerchantProfileValidator implements DotValidator {
   public messages = {}
 }
 
-/*
- * UpdateMerchantProfileValidator 
+/**
+ * UpdateMerchantProfileValidator
  */
 export class UpdateMerchantProfileValidator implements DotValidator {
   constructor(protected ctx: HttpContextContract) {
@@ -28,19 +28,19 @@ export class UpdateMerchantProfileValidator implements DotValidator {
       id: schema.string({ trim: true }, [
         rules.required(),
         rules.minLength(14),
-        rules.exists({ table: 'accounts', column: 'id', where: { user_id: this.ctx.auth.user!.id } }),
+        rules.exists({ table: 'merchant_profiles', column: 'id'}),
       ]),
     }),
     name: schema.string.optional(),
     description: schema.string.optional(),
     // type: schema.string.optional(),
-    // avatar is optional, but if it is provided, it must be a valid image
-    avatar: schema.file.optional(),
+    // photo is optional, but if it is provided, it must be a valid image
+    photo: schema.file.optional(),
   })
 
   public messages = {}
 }
-/*
+/**
  * Show MerchantProfile Validator
  *
  */
@@ -53,14 +53,17 @@ export class ShowMerchantProfileValidator implements DotValidator {
       id: schema.string({ trim: true }, [
         rules.required(),
         rules.minLength(14),
-        // rules.exists({ table: 'merchant_profiles', column: 'id', where: { account_id: this.ctx.auth.user!.id } }),
+        rules.exists({ table: 'merchant_profiles', column: 'id' }),
       ]),
     }),
+    load: schema.array.optional().members(
+      schema.enum.optional(["address", "stores", "account", "phones", "emails"] as const)
+    ),
   })
   public messages = {}
 }
-/*
- * Destroy MerchantProfile Validator 
+/**
+ * Destroy MerchantProfile Validator
  */
 export class DestroyMerchantProfileValidator implements DotValidator {
   constructor(protected ctx: HttpContextContract) {
@@ -71,7 +74,7 @@ export class DestroyMerchantProfileValidator implements DotValidator {
       id: schema.string({ trim: true }, [
         rules.required(),
         rules.minLength(14),
-        rules.exists({ table: 'accounts', column: 'id', where: { user_id: this.ctx.auth.user!.id } }),
+        rules.exists({ table: 'merchant_profiles', column: 'id', }),
       ]),
     }),
   })
@@ -79,7 +82,7 @@ export class DestroyMerchantProfileValidator implements DotValidator {
   public messages = {}
 }
 
-/*
+/**
  * List MerchantProfiles Validator
  * Handles the following:
  * - Pagination
@@ -87,32 +90,32 @@ export class DestroyMerchantProfileValidator implements DotValidator {
  * - Filtering
  * - Searching
  * - Filtering
- * 
+ *
  */
 export class ListMerchantProfilesValidator implements DotValidator {
   constructor(protected ctx: HttpContextContract) {
 
   }
   public schema = schema.create({
-    // query: schema.object().members({
-      page: schema.number.optional([
-        rules.range(1, Infinity),
-      ]),
-      limit: schema.number.optional([
-        rules.range(1, 24),
-      ]),
-      sort: schema.enum.optional(['name', 'type', 'created_at', 'updated_at'] as const),
-      order: schema.enum.optional(["asc", "desc"] as const),
-      // type: schema.enum.optional(["personal', 'business"] as const),
-      search: schema.string.optional([rules.minLength(1),]),
-      // filter: schema.string.optional({}, [
-      //   rules.minLength(1),
-      // ]),
-      load: schema.array.optional().members(
-        schema.enum.optional(["address", "stores", "account",] as const)
-      ),
+    page: schema.number.optional([
+      rules.range(1, Infinity),
+    ]),
+    limit: schema.number.optional([
+      rules.range(1, 24),
+    ]),
+    sort: schema.enum.optional(['created_at', 'updated_at'] as const),
+    order: schema.enum.optional(["asc", "desc"] as const),
+    search: schema.string.optional([rules.minLength(1),]),
+    search_by: schema.array.optional([rules.requiredIfExists('search')]).members(
+      schema.enum([] as const)
+    ),
+    load: schema.array.optional().members(
+      schema.enum.optional(["address", "stores", "account", "phones", "emails"] as const)
+    ),
+    where: schema.object.optional().members({
+      account_id: schema.string.optional(),
+    })
 
-    // }),
   })
 
   public messages = {}
