@@ -1,5 +1,6 @@
 declare module '@ioc:Adonis/Core/Validator' {
     interface Rules {
-        // uniquePhone(options: DbRowCheckOptions): Rule
+        // authUserPhoneExists(options: DbRowCheckOptions): Rule
+      authUserPhoneExists(): Rule
     }
 }
