@@ -5,6 +5,6 @@ export default class Emails extends ContactOptionSchema {
   protected tableName = 'emails'
   public setup(table: Knex.CreateTableBuilder): void {
     super.setup(table);
-    table.string('value').notNullable()//.unique()
+    table.string('value').notNullable().unique()
   }
 }
