@@ -5,6 +5,9 @@ export default class Phones extends ContactOptionSchema {
   protected tableName = 'phones'
   public setup(table: Knex.CreateTableBuilder): void {
     super.setup(table);
-    table.string('value', 12).notNullable()//.unique()
+    // table.integer('code', 3).notNullable().unique()
+    // table.integer('number', 9).notNullable().unique()
+    table.string('value', 12).notNullable().unique()
+    // table.unique(['code', 'number'])
   }
 }
