@@ -6,15 +6,16 @@ export default class ContactOptionSchema extends DotBaseSchema {
   public useValidatedAt: boolean = true;
   public useUserRelation: boolean = true;
   public useRelatedTo: boolean = true;
-  public useTimestamps: boolean = false;
+  public useTimestamps: boolean = true;
+  public usePivotTable = true
   public setup(_table: Knex.CreateTableBuilder): void {
     // the max length of email is 254
     // table.string('value').notNullable().unique()
     // the type of contact option [email, phone, etc]
     // table.string('type').notNullable()
     // table
-    //   .string('related_to')
+    //   .string('related_id')
     //   .references('users.id')
-    //   .onDelete('CASCADE') // delete when user is deleted 
+    //   .onDelete('CASCADE') // delete when user is deleted
   }
 }
