@@ -6,9 +6,26 @@ Route.group(() => {
     Route.group(() => {
         require("./v1/auth/users")
     }).prefix("auth/users")
+
+    //
     Route.group(() => {
         require("./v1/accounts")
-    })//.prefix("")
+    })
+    Route.group(() => {
+        require("./v1/merchant_profiles")
+    })
+    Route.group(() => {
+        require("./v1/customer_profiles")
+    })
+    Route.group(() => {
+        require("./v1/stores")
+    })
+    Route.group(() => {
+        require("./v1/categories")
+    })
+    Route.group(() => {
+      require("./v1/sections")
+    })
 
 
 }).prefix("api/v1")
