@@ -1,26 +1,54 @@
+/*
+|--------------------------------------------------------------------------
+| Preloaded File
+|--------------------------------------------------------------------------
+|
+| Any code written inside this file will be executed during the application
+| boot.
+|
+*/
 import { string } from '@ioc:Adonis/Core/Helpers'
 import { validator } from '@ioc:Adonis/Core/Validator'
 import Phone from 'App/Models/ContactOptions/Phone'
+import { AuthPivotTags } from 'App/Models/User'
 
-// validator.rule('uniquePhone', async (value, params, options) => {
-//     if (typeof value !== 'string') {
-//         return
-//     }
-//     // for each phone number in table of phones
-//     // check if rolatedType is the same as the current model
-//     // if so, check if the phone number is the same
-//     // if so, return false
-//     // if not, return true
-//     const phones = await Phone.query()
-//         .where('relatedTo', 'User')
-//         .where('number', value)
-//         .first()
-//     if (phones) {
-//         options.errorReporter.report(
-//             options.pointer,
-//             'uniquePhone',
-//             'uniquePhone validation failed',
-//             options.arrayExpressionPointer
-//         )
-//     }
-// })
+validator.rule('authUserPhoneExists', async (value, params, options) => {
+  const phone = await Phone.query()
+    .where('value', value)
+    .first()
+  var failed = false
+  if (phone) {
+    const pivot = await phone.related('users').pivotQuery().where('tag', AuthPivotTags.user).first()
+    if (!pivot) {
+      failed = true;
+    }
+  } else {
+    failed = true;
+  }
+  if (failed) {
+    options.errorReporter.report(
+      options.pointer,
+      'authUserPhone',
+      'authUserPhone validation failed',
+      options.arrayExpressionPointer
+    )
+  }
+  // options.pointer = "options.pointer"
+  // options.arrayExpressionPointer = "options.arrayExpressionPointer"
+  // options.tip = "options.tip"
+  // options.field = "options.field"
+  // options.mutate = (value) => {
+  //   return "options.mutate"
+  // }
+
+},
+  (options, type, subtype) => {
+    if (subtype !== 'string') {
+      throw new Error('"authUserPhoneExists" rule can only be used with a string schema type')
+    }
+
+    return {
+      async: true,
+    }
+  }
+)
