@@ -49,8 +49,8 @@ export const reporters: Config['reporters'] = [specReporter()]
 export const runnerHooks: Required<Pick<Config, 'setup' | 'teardown'>> = {
   setup: [
     () => TestUtils.ace().loadCommands(),
-    () => TestUtils.db().migrate(),
-    () => TestUtils.db().seed()
+    // () => TestUtils.db().migrate(),
+    // () => TestUtils.db().seed()
   ],
   teardown: [],
 }
