@@ -7,9 +7,9 @@ import 'sample_item_details_view.dart';
 /// Displays a list of SampleItems.
 class SampleItemListView extends StatelessWidget {
   const SampleItemListView({
-    Key? key,
+    super.key,
     this.items = const [SampleItem(1), SampleItem(2), SampleItem(3)],
-  }) : super(key: key);
+  });
 
   static const routeName = '/';
 
